- fast result for local search in case that less than 10 hits exists - small change in display of RAM in profiling
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@6326 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/search/RankingProcess.java
@@ -358,7 +358,7 @@ public final class RankingProcess extends Thread {
         return bestEntry;
     }
     
-    public URLMetadataRow takeURL(final boolean skipDoubleDom, final long timeout) {
+    public URLMetadataRow takeURL(final boolean skipDoubleDom, final int timeout) {
         // returns from the current RWI list the best URL entry and removes this entry from the list
     	long timeLimit = System.currentTimeMillis() + timeout;
     	while (System.currentTimeMillis() < timeLimit) {
